+++ CHANGELOG.md
@@ -651,7 +651,7 @@ This is a bugfix release.
 * Strong's support ([#10][i10])
   - Strong's transcriptions/original word are displayed as little hint/pop-up above the word when hovering the mouse over the corresponding word.
   - Detailed Strong's information is shown on the bottom left in a new area below the tags list.
-  - Strong's based search with link 'Find all occurrances' in info box.
+  - Strong's based search with link 'Find all occurrences' in info box.
 * Dutch translation (Thanks to Tom Lemmens)
 
 ### Enhancements